pyproject.toml
[tool.poetry]
name = "goit-web-module11"
version = "0.1.0"
description = ""
authors = ["Nataliia Tiutiunnyk <nat.tiutiunnyk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
dotenv = "^0.0.5"
asyncpg = "^0.29.0"
email-validator = "1.3.1"
phonenumbers = "^8.13.30"
libgravatar = "^1.0.4"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
fastapi-jwt-auth = "0.5.0"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.1"
redis = "^5.0.2"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.39.0"
fastapi-mail = "1.2.5"
httpx = "^0.27.0"
pytest-cov = "^4.1.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
